backend/requirements.txt
# Framework e Servidor
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings

# Banco de Dados e ORM
SQLAlchemy[asyncio]
psycopg2-binary

# Segurança e Autenticação
passlib==1.7.4
bcrypt==3.2.2
python-jose[cryptography]

# Ferramentas de Utilidade
typer 
pandas
openpyxl
email-validator
python-multipart
httpx
twilio
# Inteligência Artificial
google-generativeai